reorganizing and gui work
